requirements.txt
annotated-types==0.6.0
anyio==3.7.1
certifi==2023.7.22
click==8.1.7
colorama==0.4.6
fastapi==0.104.1
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
iniconfig==2.0.0
mslex==0.3.0
packaging==23.2
pluggy==1.3.0
psutil==5.9.6
psycopg2-binary==2.9.9
pydantic==2.5.1
pydantic_core==2.14.3
pytest==7.4.3
python-dotenv==1.0.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
taskipy==1.12.0
tomli==2.0.1
typing_extensions==4.8.0
uvicorn==0.24.0.post1
